+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased - 2021-xx-xx
+- Add `DisplayStream` responder.
 
 
 ## 0.5.0 - 2022-01-18